rearranged local and global TC
