+++ cli/src/main/java/com/devonfw/tools/ide/tool/eclipse/Eclipse.java
@@ -53,7 +53,7 @@ public class Eclipse extends IdeToolCommandlet {
    */
   protected ProcessResult runEclipse(boolean log, String... args) {
 
-    Path toolPath = Paths.get(getBinaryName());
+    Path toolPath = Paths.get(getName());
     ProcessContext pc = this.context.newProcess();
     if (log) {
       pc.errorHandling(ProcessErrorHandling.ERROR);
